pyproject.toml
[tool.poetry]
name = "print-service"
version = "0.1.0"
description = ""
authors = ["Inasu <just.me.juncy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
celery = "^5.4.0"
django-cors-headers = "^4.3.1"
requests = "^2.32.3"
logging = "^0.4.9.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120